硬件数据数据json落盘：stars_soc和stars_chip_trans
# message auto-generated for no-merge-commit merge:
merge develop into master

硬件数据数据json落盘：stars_soc和stars_chip_trans

Created-by: msprofDeveloper

Match-id-50c9dd0fcae6c8d7b00df1aa545c78403b155644

+++ analysis/csrc/domain/entities/json_trace/include/trace_event.h
@@ -67,14 +67,14 @@ private:
 class CounterEvent : public TraceEvent {
 public:
     CounterEvent(uint32_t pid, int tid, const std::string &ts, const std::string &name);
-    void SetSeriesValue(const std::string &key, const uint64_t &value);
+    void SetSeriesValue(const std::string &key, const double &value);
 private:
     void ToJson(JsonWriter &ostream) override;
     void ProcessArgs(JsonWriter &ostream);
 private:
     std::string ts_;
     std::string ph_ = "C";
-    std::unordered_map<std::string, uint64_t> seriesValue_;
+    std::unordered_map<std::string, double> seriesValue_;
 };
 
 /**
